fix throttle test for key queues
