+++ pkg/execution/state/redis_state/lua/includes/gcra.lua
@@ -47,7 +47,7 @@ local function gcraUpdate(key, now_ms, period_ms, limit, burst, capacity)
   -- calculate next theoretical arrival time
   local new_tat = tat + (math.max(capacity, 1) * emission)
 
-  if capacity >= 0 then
+  if capacity > 0 then
     local expiry = (period_ms / 1000)
     redis.call("SET", key, new_tat, "EX", expiry)
   end
