+++ pkg/inngest/function.go
@@ -177,9 +177,9 @@ func (t *Throttle) UnmarshalJSON(in []byte) error {
 	if t.Limit == 0 {
 		t.Limit = 1
 	}
-	if t.Burst == 0 {
-		t.Burst = 1
-	}
+	// if t.Burst == 0 {
+	// 	t.Burst = 1
+	// }
 	return err
 }
